完成编码器测试

+++ Project_Headers/encoder.h
@@ -14,9 +14,9 @@ typedef struct{
 	EMIOSn_CH ch;   	//电磁组采用的EMIOS通道为具有MCB功能的ch16
 	float _transmission_ratio;// 传动比 0.16515*29/28 = 0.17105// 新车0.15502
 	float _radius;		//车轮半径 32mm？
-	float _speed;		//车轮相对于地面的绝对速度
+	int _speed;		//车轮相对于地面的绝对速度
 	float _resolution; 	//光编线数
-	uint32_t _last_counter;
+	int _last_counter;
 	float _time;		//采集周期
 	uint8_t dir;		//车速的方向
 	uint8_t dir_pad;
